Format effect descriptions by displaying the 'label' or 'target' identifier
Fixes #30

This addition searches for a pattern - [label]{category:target} - where the label is the text to be displayed (sometimes could be ""), the category can be item, move, ability, pokemon, etc., and the target is the identifier of the object. If the label text is empty (equals ""), then the target identifier is used, for now (although the object name attribute should really be used here).

Ideally, this should be modified in the future so that the text displayed is clickable and links to the appropriate detail page.