Load Pokémon's moves when fragment becomes visible
When the fragment becomes visible, the data is loaded. However, when using the tab buttons as opposed to swiping, we need to check if getActivity() is null, because it can give an NPE. This means learnset data is not loaded automatically when using the tab buttons to navigate more than one tab away.

Resolves #41